refactor: 로그인, 탈퇴 페이지 dynamic route를 통한 OAuth Provider 참조로 변경 (#28)
