+++ src/app/delete/[OAuthProvider]/page.tsx
@@ -3,7 +3,7 @@
 import { Suspense } from 'react';
 import HeaderLayout from '@/components/layout/header-layout';
 import { Button } from '@/components/ui/button';
-import DeleteContent from './DeleteContent';
+import DeleteContent from './content';
 
 export default function Page() {
   const handleDeleteOAuthAccount = () => {
@@ -18,7 +18,7 @@ export default function Page() {
           삭제
         </Button>
         <Suspense fallback={<div>로딩중...</div>}>
-          <DeleteContent />
+          <DeleteContent OAuthProvider="KAKAO" />
         </Suspense>
       </div>
     </HeaderLayout>
